+++ src/main/java/roomescape/controller/rest/TimeController.java
@@ -1,6 +1,8 @@
-package roomescape.controller;
+package roomescape.controller.rest;
 
 import jakarta.validation.Valid;
+import java.net.URI;
+import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,9 +14,6 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.time.TimeRequest;
 import roomescape.dto.time.TimeResponse;
 import roomescape.service.TimeService;
-
-import java.net.URI;
-import java.util.List;
 
 @RestController
 @RequestMapping("/times")
